chore: move css files for better organization
