+++ src/app/(pages)/project/2025/components/SpeakerGrid.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import styles from "../Cards.module.css";
+import styles from "../css/Cards.module.css";
 
 import { Box, Title, Text } from "@mantine/core";
 import SpeakerCard from "./SpeakerCard";
